codebase-assitant/requirements.txt
fastapi
uvicorn
pydantic

langchain
langchain-community
langchain-huggingface

langchain-groq

faiss-cpu

sentence-transformers
transformers
torch

gitpython

python-dotenv
tqdm

python-multipart
chardet
pydantic
qdrant-client
langchain-qdrant